Fix circular deps
- assay and entities sub packages were depended on by main package
- AppContext had circular dep with entities/SampleTypeAppContext and assay/AssayAppContext
