+++ packages/components/src/index.ts
@@ -1485,7 +1485,7 @@ export type { UsersLoader } from './internal/components/forms/actions';
 export type { LineageGroupingOptions } from './internal/components/lineage/types';
 export type { AnnouncementModel, ThreadActions } from './internal/announcements/model';
 export type { AnnouncementsAPIWrapper } from './internal/announcements/APIWrapper';
-export type { AdminAppContext, AppContext, ExtendableAppContext } from './internal/AppContext';
+export type { AdminAppContext, AppContext, AssayAppContext, ExtendableAppContext, SampleTypeAppContext } from './internal/AppContext';
 export type { WithAdminAppContext } from './internal/components/administration/useAdminAppContext';
 export type { ThreadBlockProps } from './internal/announcements/ThreadBlock';
 export type { ThreadEditorProps } from './internal/announcements/ThreadEditor';
@@ -1499,3 +1499,4 @@ export type {
     LabelPrintingProviderProps,
     LabelPrintingContextProps,
 } from './internal/components/labels/LabelPrintingContextProvider';
+export type { SamplesEditableGridProps } from './internal/sampleModels';
